moved the main file ipc handlers into an ipc handlers subfolder
