+++ src/Screens/EditNoteScreen.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
+const EditNoteScreen = () => {
+  const navigate = useNavigate();
+  return (
+    <div>
+      <button onClick={() => navigate('/')}>Go Back To Initial Screen</button>
+      <h1>Alright This screen will be used for Editing Notes</h1>
+    </div>
+  );
+};
+
+export default EditNoteScreen;
