+++ src/Screens/InitialScreen.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import NoteUploader from '../Components/NoteUploader.jsx';
 import SexyImage from '../Components/SexyImage.jsx';
 import MapRepoFolder from '../Components/MapRepoFolder.jsx';
@@ -8,8 +8,13 @@ const InitialScreen = () => {
   const navigate = useNavigate();
   return (
     <div style={{ padding: 16 }}>
-      <button onClick={() => navigate('/noteuploadscreen')}>GoToNotes</button>
-      <button onClick={() => navigate('/imageuploadscreen')}>GoToImages</button>
+      <button onClick={() => navigate('/noteuploadscreen')}>
+        Upload Notes
+      </button>
+      <button onClick={() => navigate('/imageuploadscreen')}>
+        Upload Image
+      </button>
+      <button onClick={() => navigate('/editnotescreen')}>Edit Notes</button>
       <Products />
       <NoteUploader />
       <MapRepoFolder />
