+++ src/index.html
@@ -2,8 +2,7 @@
 <html>
   <head>
     <meta charset="UTF-8" />
-
-    <title>ZFileUploader</title>
+    <title>DegenerateUploader</title>
   </head>
   <body>
     <div id="root"></div>
